See merge request oai/openairinterface5g!2459

* !2419 chore(ci): running 5G RF simulator scenarios with address sanitizer
* !2455 Remove mac_LogicalChannelConfig from RLC (handled at MAC)
* !2456 Setting uniform periodicity for CSI-RS and measurement report
* !2429 LDPC encoding/decoding offload to the T2 card
* !2447 E2 agent: build in CI, fix slice support, add gNB-DU and gNB-CU-UP IDs
* !2452 Include sample advance option in UE tutorial
* !2430 bugfix: nr-ue: set first_tx=1 at first transmission of a harq process
* !2457 Fix scaling for FR2

+++ openair3/UTILS/conversions.h
@@ -110,10 +110,10 @@ do {                            \
 /* Convert an array of char containing vALUE to x */
 #define BUFFER_TO_INT32(buf, x) \
 do {                            \
-    x = ((buf)[0] << 24) |      \
-        ((buf)[1] << 16) |      \
-        ((buf)[2] << 8)  |      \
-        ((buf)[3]);             \
+    x = (((uint32_t)(buf)[0]) << 24) |      \
+        (((uint32_t)(buf)[1]) << 16) |      \
+        (((uint32_t)(buf)[2]) << 8)  |      \
+        (((uint32_t)(buf)[3]));             \
 } while(0)
 
 /* Convert an array of char containing vALUE to x */
@@ -427,10 +427,10 @@ do {                                                    \
 do {                                                                    \
     DevCheck((bITsTRING)->size == 4, (bITsTRING)->size, 4, 0);          \
     DevCheck((bITsTRING)->bits_unused == 0, (bITsTRING)->bits_unused, 0, 0); \
-    mACRO = ((bITsTRING)->buf[3] << 24) +                               \
-            ((bITsTRING)->buf[2] << 16) +                               \
-            ((bITsTRING)->buf[1] << 8) +                                \
-            ((bITsTRING)->buf[0]);                                      \
+    mACRO = (((uint32_t) (bITsTRING)->buf[3]) << 24) +                               \
+            (((uint32_t) (bITsTRING)->buf[2]) << 16) +                               \
+            (((uint32_t) (bITsTRING)->buf[1]) << 8) +                                \
+            (((uint32_t) (bITsTRING)->buf[0]));                                      \
 } while (0)
